Query kinase->substrate rules from corpus
Changed query in get_statements to focus specifically on statements of
the type where a kinase protein abundance directy increases the
abundance of a modified protein (i.e., a phosphorylation event).

The get_monomers function also includes sites/states to specify the
activities of monomers (e.g., to indicate that a kinase is active or
inactive).

Added code to handle cases where identifiers start with numbers, by
prepending underscore (e.g., '14_3_3_Family' now becomes '_14_3_3_Family'
so that this is a valid Python identifier).