Make binding rules from complex statements
Very simple right now--simple adds a single binding site, 'b', for any
protein that participates in a binding rule, and then creates a bond
between these b sites for all members participating in a complex (a
"star" conformation).

A better version would be to create a named site for every binding
partner, and have the binding between proteins use that site, e.g.
A(b=1) % B(a=1).