Determine which statements are unconverted
Wrote a function, get_all_direct_statements, that queries for all of the
statements in the corpus that have a directInc/Dec relationship, and
returns them. Also modified the other funcs to return the statements
that they have converted. The script at the end compares the two lists
to identify types of statements that have not yet been converted.

This might not be something worth keeping around (especially not in this
hacky form) but it's certainly useful to understand the conversion
process better.