Switch to storing query results as Statement objects
The idea is that there should be an intermediate data structure for
storing the relevant information from queries to the RDF corpus, but
that doesn't yet involve converting them to fully mechanistic PySB
expressions. This allows the intermediate representation to be queried
during the processing process (e.g., processing later statements can look for
statements of certain type, like activating mutations, without losing
the semantics of these statements by converting them to flat rules).

In addition, this allows the graph to be queried once, stored as belpy
statements, and then converted into PySB models in a variety of ways
without having to requery the graph.