Added activity/activity implementation.
Currently only gets directlyIncreases rules, though directlyDecreases
would be trivial to implement. Note however, that an inhibitory rule of
this type would have no effect on a substrate that had an activating
modification that had been added by another enzyme. An attempt to simply
reverse the activity would be countered by the action of a rule that
would reactivate it based on the modification.