Query for evidence, citation, annotations
This required an additional set of queries for every statement. Two
queries are actually required, because it appears that though there is
only ever one evidence text or citation for a statement, there can be
several citations (tissue, cell, species, etc.).

In addition, this required modifying all statement constructor calls to
pass this information through.

Finally, I also modified oneliner.bel to include some of this metadata
since it's useful for testing.