Get all complexes in corpus and assemble with simple convention
Current convention is to allow simultaneous binding between all members
of a complex, with named sites for each binding partner. Seems like the
minimal ("maximum entropy", in Walter's words) assumption to make, if it
is almost assuredly non-physical.

The current issue is that with this implementation all bonds in the
complex share the same bond index. This will be fixed in the next
commit.