Refactor rdf_to_pysb core into a func
This way it can be called by other scripts on different files to make
models which can subsequently be tweaked for analysis, etc.

Also note that the abbrevs and states dicts were moved to statements.py,
because that's really where they belonged. That's where the information
was used, and the previous way actually led to a circular import which
caused problems.