Initial impl of interactions_only assembly for Phosphorylations
The interactions_only "policy" specifies only the binding relationships
between proteins, which can subsequently be used to generate a contact
map which is useful for understanding the overall structure of the
network.

Phosphorylations are represented by a binding between the modified site
of the substrate and the kinase_site (active site) of the kinase.