Moved policy checking code to Statement superclass
This means that each statement type implements the assembly/monomer
methods for the different policies, without having to duplicate the
if/then checking code each time.

Note that this only works well for "global" policy types, where the
policy is not specific to any particular type of statement. This
approach will have to be revisited when the policy specification becomes
more complex, allowing for more granular control over the assembly of
specific types of statements.