Proof-of-concept model checking using influence map
Identifies rules in the model that have the input node among the
reactant patterns, and output rules that have the output nodes among the
product patterns. Then checks the influence map to see if there is a
path in the influence map between these rules.

This could also be done using the species_graph or reaction network, but
this would not be robust to combinatorially complex models.