Path polarity computation
Needed because even though the influence map has directional edges, many
of these edges are actually inhibitory (i.e., the firing of one rule
actually diminishes the activity of another rule). So it is necessary to
compute the polarity of the overall path and check if it matches the
constraint.