Allow renaming to be done at mapping time
Avoids looking up incorrectly grounded non-human genes in Uniprot, which
is very costly. The alternative would be to 1) Add the naming field
explicitly into the grounding map, or 2) add an ad-hoc field to
statements specifying that they had been mapped, or 3) separately
returning lists of mapped and unmapped statements and having the caller
rename only those that are mapped. This seemed to be the easiest thing.