Build protein map from texts with grounding.
This function looks at the grounding of the entity texts extracted from
the statements and finds proteins where there is grounding to a human
protein that maps to an HGNC name that is an exact match to the entity
text. Returns a dict that can be used to update/expand the grounding
map.