Fix for ElementTree parser issue
The issue is that ASCII only strings are returned by the
ElementTree.parse method as str (bytestrings) rather than unicode
in Python 2. This makes it hard to test functions for maintaining
unicode within the program. Wrapping str() (after importing builtins
... str) doesn't work because it turns Nones into 'None'; similarly
calling decode requires explicitly testing every returned object to
see if it is None. This fix was found from a message board and involves
subclassing the XMLTreeBuilder class and overriding a single method. The
trick is how to use the same syntax (for passing in an instance of this
parser) to Python 3. The trickery is to create a class,
UnicodeXMLTreeBuilder, in the case of Python 2, and a function with the
same name in Python 3. In Python 2, parser=UnicodeXMLTreeBuilder()
returns an instance of the subclassed parser, whereas in Python 3, it
calls a function that returns, causing the system to use the default
parser, which is the desired behavior.