Fixed binning bug; limited bins [3:9]
A binning bug existed where the bin edges retrieved by histogram
included all bin edges, not just the right edge as previously thought

edge    edge    edge    edge
 0       1       2       3
 | bin 1 | bin 2 | bin 3 |

 code was just checking if expression value < edge value
 sometimes it binned values into bin+1
 Got rid of first edge, works now.

In addition, limited n_bins to be between 3 and 9. Although our colorscales
can technically map to any number of bins, these get harder to discriminate.
Furthermore, this should even let us make the color scales more pleasing as
the ones on colorbrewer2.com go from 3-9 and do not map the a 3 bin scale to
the same color range as a 9 bin scale. Binning a 9 bin scale to the same range
as that of a 3 bin would make the bins indiscernable, while mapping a 3 bin to
the same range as a 9 bin makes the scale ugly...