Update preassembler to fix bug
In this version, groups are created based on an entity_matches_key when
an agent is not part of a hierarchy component. This ensures that no
statements that are not part of the hierarchy will fall through the
cracks.

It's possible that the elimination of the separate preassembly
procedures for the statements with no components will slow things down
slightly, so this might be worth looking into.