Renaming nodes and columns
More rules added to renaming nodes (SIF) and columns (MIDAS).
CNO doesn't tolerate 'AND', '-', or things starting with a digit.
These are replaced with 'A_ND', '_', and by prepending 'abc_', respectively.
The replacements for SIF must be done after a list of agent names is made
for MIDAS filtering. Otherwise, anything that is renamed won't be included
in the MIDAS file. Columns in the MIDAS file are subjected to the same
replacement rules once the dataframe is generated.