Seperate functions to save js and json
Seperate functions were necessary to save js and json. It is impractical
to load a js file after the page has loaded and this is necessary for
having interactivity when it comes to loading data  with a rendering of the
pathway map.

The old version of save_model is retained as it is still used by the notebook.
The new version of print_cyjs() generates a json string for save_json().
save_model() does not rely on this string anymore.