Enable CORS
The python API is accessible using python's requests library, however, it was
not accessible using jQuery AJAX calls. This is a security feature of most
modern browsers which expect a header to be returned by the API, in addition
to the JSON.

I have enabled this for the reach text parser and the cyjs assembler.
Currently, CORS is enabled for all origins, however, this might not be a great
idea in production.

https://developer.mozilla.org/en-US/docs/Web/HTTP/Access_control_CORS
https://mushfiq.me/2015/10/07/enable-cors-in-bottle-python/
http://stackoverflow.com/questions/9713644/when-is-it-safe-to-enable-cors