Sadly, cycle elimination is not going to work
Cycles cannot be eliminated during the generation of the reachable sets.
Consider the graph

   0
   |\
   | 1
   |/
   2
  /|
 3 |
  \|
   4

where all arrows flow downward. Reachable at 0: is 0; reachable at 1 is
(1, 2); reachable at 2 is (2, 3). If at this point we remove 2 from the
reachable set because it's already been visited, then we will omit a
path of length 4 that goes through 1, and will only get the path
[0, 2, 3, 4].