More tests, including failing test of sampling
Error in test_sampling is IndexError: Cannot choose from an empty sequence

Not sure if it is permissible for sampling to arrive at a node where
there is no valid path to the target (no successors of v containing tags
associated with all elements in the current path). If this is
permissible, then the path needs to be discarded. If not, then there is
a problem with the calculation of the dictionary.