Re-tool _clear functionality in DatabaseManager.
Instead of checking for the name 'primary' when dropping tables,
a new kwarg, force, is not passed to _clear and drop_tables
which, if False (the default), brings up the warning message on
drop_tables. Also enable a selective list of tables to be dropped
or created. This is now used by the upload procedure, and the
force paramter set as true in all tests, where appropriate.