Add BIR1 and Bir1 to misgrounding for KCNJ6
This is probably correct sometimes, but overwhelmingly BIR1 refers to
the yeast protein BAK1 interacting receptor like kinase 1. Adding to
misgrounding for now. It's difficult to get sufficient training data
because only a small number of texts have a defining pattern for BIR1.
If needed, we could handle this case by manually annotating a training corpus