Fix data schema of annotations
In PyBEL v14, the annotations data structure should be a `Mapping[str, Mapping[str, bool]]` where the first key is the name of the annotation, the nested key is the value of the annotation, and the bool is always just True (unless you have some negative annotations, then set it to false!). Even for annotations that only have a singleton value and the positive/negative dichotomy is irrelevant, it should still be done this way. Typically, even developers won't be accessing the networkx Graph class's `add_edge()` function and the PyBEL wrappers on it should handle of a variety of data structures, such as this one, but since we're accessing it directly in the PyBEL assembler, this is a problem.

Addresses https://github.com/sorgerlab/indra/issues/1184#issuecomment-718036895