Refactor instantiation of agents
The code for standardizing the db_refs and name showed up in many places, so i refactored it into a classmethod for the Agent class. The only caveat is that the standardize_name_db_refs function is imported inside since it would cause circular dependencies.

Maybe this function needs a different name though, since it has the same signature as Agent.__init__() and I don't want to cause confusion