Add DGI Processor
Closes #1259

This PR adds a processor for the DGI database. It automates acquisition of the data and stores via PyStow. It applies a very opinionated curation of the interaction types annotated in the database to decide which INDRA statement type each drug-target interaction should be, with several conflicts to the philosophy of the DrugBank converter (notably, agonist, antagonist, and inverse agonist can't be used to determine the polarity of the relationship, only that it has the same/opposite effect as the native ligand).

It produces 11534 statement, unfortunately throwing away many apolar statements we could keep if it were possible to instantiate the RegulatesActivity class.

This PR also brings over some of the functionality introduced in #1261 for instantiation of agents, so it would be nice to review that one first.